moving back bipartite impl to cc
